server/requirements-dev.txt
pytest>=9,<10
httpx>=0.27,<1
PyJWT>=2.10,<3
cryptography>=45,<47
redis>=5,<7
fakeredis>=2.31,<3
opentelemetry-api>=1.28,<2
opentelemetry-sdk>=1.28,<2
opentelemetry-exporter-otlp-proto-http>=1.28,<2
